changed type into methods of controller

+++ src/main/java/com/car/rental/mycarrental/service/BodyServiceImpl.java
@@ -3,6 +3,7 @@ package com.car.rental.mycarrental.service;
 
 import com.car.rental.mycarrental.dao.BodyRepository;
 import com.car.rental.mycarrental.entity.Body;
+import com.car.rental.mycarrental.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,28 +17,36 @@ public class BodyServiceImpl implements BodyService {
 
 
     @Override
-    public List<Body> getAllBodies() {
-
+    public List<Body> getBodies() {
         return bodyRepository.findAll();
     }
 
     @Override
-    public void saveBody(Body body) {
-        bodyRepository.save(body);
+    public Optional<Body> getBodyById(Integer id) {
+        if(!bodyRepository.existsById(id))
+            throw new NotFoundException(id);
+        return bodyRepository.findById(id);
     }
 
     @Override
-    public Body getBody(int id) {
-        Body body = null;
-        Optional<Body> newbody = bodyRepository.findById(id);
-        if (newbody.isPresent()) {
-            body = newbody.get();
-        }
-        return body;
+    public Body saveBody(Body body) { return bodyRepository.save(body);
     }
+
+    @Override
+    public void updateBody(Body body, Integer id) {
+        if(!bodyRepository.existsById(id))
+            throw new NotFoundException(id);
+
+        Body current = bodyRepository.getById(id);
+        current.updateBody(body);
+        bodyRepository.save(current);
+    }
+
 
     @Override
     public void deleteBody(Integer id) {
+        if(!bodyRepository.existsById(id))
+            throw new NotFoundException(id);
         bodyRepository.deleteById(id);
     }
 }
