+++ src/main/java/com/car/rental/mycarrental/service/EmployeeServiceImpl.java
@@ -3,6 +3,7 @@ package com.car.rental.mycarrental.service;
 
 import com.car.rental.mycarrental.dao.EmployeeRepository;
 import com.car.rental.mycarrental.entity.Employee;
+import com.car.rental.mycarrental.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,27 +17,36 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 
     @Override
-    public List<Employee> getAllEmployees() {
+    public List<Employee> getEmployees() {
         return employeeRepository.findAll();
     }
 
     @Override
-    public void saveEmployee(Employee employee) {
-        employeeRepository.save(employee);
+    public Optional<Employee> getEmployeeById(Integer id) {
+       if(!employeeRepository.existsById(id))
+           throw new NotFoundException(id);
+       return employeeRepository.findById(id);
     }
 
     @Override
-    public Employee getEmployee(int id) {
-        Employee employee = null;
-        Optional<Employee> newEmployee = employeeRepository.findById(id);
-        if(newEmployee.isPresent()){
-            employee = newEmployee.get();
-        }
-        return employee;
+    public Employee saveEmployee(Employee employee) {
+        return employeeRepository.save(employee);
+    }
+
+    @Override
+    public void updateEmployee(Employee employee, Integer id) {
+        if(!employeeRepository.existsById(id))
+            throw new NotFoundException(id);
+
+        Employee current = employeeRepository.getById(id);
+        current.updateEmployee(employee);
+        employeeRepository.save(current);
     }
 
     @Override
     public void deleteEmployee(int id) {
+        if(!employeeRepository.existsById(id))
+            throw new NotFoundException(id);
         employeeRepository.deleteById(id);
     }
 }
